requirements.txt
fastapi>=0.95,<0.100
uvicorn>=0.22,<0.24
sqlalchemy>=2.0,<2.1
pydantic>=1.10,<2.0
python-multipart>=0.0.6,<0.1
jinja2>=3.1,<3.2
email-validator>=2.0,<3
python-dotenv>=1.0,<2
openpyxl>=3.1,<3.2
